+++ src/actor_state.rs
@@ -1,0 +1,6 @@
+#[derive(PartialEq, Clone, Copy, Debug)]
+pub enum ActorState {
+    Running,
+    Sleeping,
+    Stopped,
+}