+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - move Actor Messages into separate directory
 - move Mailbox to separate file
 - move ActorHandler to separate file
+- move ActorState to separate file
 
 # WIP 0.0.8